pyproject.toml
[tool.poetry]
name = "jetstream.py"
version = "0.1.0"
description = ""
authors = ["charbonnierg <guillaume.charbonnier@araymond.com>"]
packages = [{include = "jsm", from="src" }]

[tool.poetry.dependencies]
python = "^3.8"
asyncio-nats-client = "^0.11.4"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.0.2"
flake8 = "^3.9.2"
black = "^21.7b0"
mypy = "^0.910"
datamodel-code-generator = "^0.11.8"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
